feat: use kotlin extensions feature to replace the unneeded messenger class (#1811)
The kotlin class extensions feature makes code easier to understand. The previous wrapper class was not needed at all.
